style: reconstruction of POST/PUT/DELETE/PATCH /api/resources, PATCH /api/paste, GET /api/raw and GET /api/preview
