+++ pkg/http/raw.go
@@ -1,233 +1,26 @@
 package http
 
 import (
-	"errors"
 	"files/pkg/common"
 	"files/pkg/drives"
 	"k8s.io/klog/v2"
 	"net/http"
-	"net/url"
-	gopath "path"
-	"path/filepath"
-	"strings"
-
-	"github.com/mholt/archiver/v3"
-
-	"files/pkg/files"
-	"files/pkg/fileutils"
+	"time"
 )
 
-func slashClean(name string) string {
-	if name == "" || name[0] != '/' {
-		name = "/" + name
-	}
-	return gopath.Clean(name)
-}
-
-func parseQueryFiles(r *http.Request, f *files.FileInfo) ([]string, error) {
-	var fileSlice []string
-	names := strings.Split(r.URL.Query().Get("files"), ",")
-
-	if len(names) == 0 {
-		fileSlice = append(fileSlice, f.Path)
-	} else {
-		for _, name := range names {
-			name, err := url.QueryUnescape(strings.Replace(name, "+", "%2B", -1))
-			if err != nil {
-				return nil, err
-			}
-
-			name = slashClean(name)
-			fileSlice = append(fileSlice, filepath.Join(f.Path, name))
-		}
-	}
-
-	return fileSlice, nil
-}
-
-func parseQueryAlgorithm(r *http.Request) (string, archiver.Writer, error) {
-	switch r.URL.Query().Get("algo") {
-	case "zip", "true", "":
-		return ".zip", archiver.NewZip(), nil
-	case "tar":
-		return ".tar", archiver.NewTar(), nil
-	case "targz":
-		return ".tar.gz", archiver.NewTarGz(), nil
-	case "tarbz2":
-		return ".tar.bz2", archiver.NewTarBz2(), nil
-	case "tarxz":
-		return ".tar.xz", archiver.NewTarXz(), nil
-	case "tarlz4":
-		return ".tar.lz4", archiver.NewTarLz4(), nil
-	case "tarsz":
-		return ".tar.sz", archiver.NewTarSz(), nil
-	default:
-		return "", nil, errors.New("format not implemented")
-	}
-}
-
-func setContentDisposition(w http.ResponseWriter, r *http.Request, file *files.FileInfo) {
-	if r.URL.Query().Get("inline") == "true" {
-		w.Header().Set("Content-Disposition", "inline")
-	} else {
-		// As per RFC6266 section 4.3
-		w.Header().Set("Content-Disposition", "attachment; filename*=utf-8''"+url.PathEscape(file.Name))
-	}
-}
-
 func rawHandler(w http.ResponseWriter, r *http.Request, d *common.Data) (int, error) {
+	start := time.Now()
+	klog.Infoln("Function rawHandler starts at", start)
+	defer func() {
+		elapsed := time.Since(start)
+		klog.Infof("Function rawHandler execution time: %v\n", elapsed)
+	}()
+
 	srcType := r.URL.Query().Get("src")
-	if srcType == "sync" {
-		return http.StatusNotImplemented, nil
-	} else if srcType == "google" {
-		bflName := r.Header.Get("X-Bfl-User")
-		src := r.URL.Path
-		metaData, err := drives.GetGoogleDriveMetadata(src, w, r)
-		if err != nil {
-			klog.Error(err)
-			return common.ErrToStatus(err), err
-		}
-		if metaData.IsDir {
-			return http.StatusNotImplemented, errors.New("doesn't support directory download for google drive now")
-		}
-		return drives.RawFileHandlerGoogle(src, w, r, metaData, bflName)
-	} else if srcType == "cloud" || srcType == "awss3" || srcType == "tencent" || srcType == "dropbox" {
-		bflName := r.Header.Get("X-Bfl-User")
-		src := r.URL.Path
-		metaData, err := drives.GetCloudDriveMetadata(src, w, r)
-		if err != nil {
-			klog.Error(err)
-			return common.ErrToStatus(err), err
-		}
-		if metaData.IsDir {
-			return http.StatusNotImplemented, errors.New("doesn't support directory download for " + srcType + " drive now")
-		}
-		return drives.RawFileHandlerCloudDrive(src, w, r, metaData, bflName)
-	}
-
-	file, err := files.NewFileInfo(files.FileOptions{
-		Fs:         files.DefaultFs,
-		Path:       r.URL.Path,
-		Modify:     true,
-		Expand:     false,
-		ReadHeader: d.Server.TypeDetectionByHeader,
-	})
+	handler, err := drives.GetResourceService(srcType)
 	if err != nil {
-		return common.ErrToStatus(err), err
+		return http.StatusBadRequest, err
 	}
 
-	if files.IsNamedPipe(file.Mode) {
-		setContentDisposition(w, r, file)
-		return 0, nil
-	}
-
-	if !file.IsDir {
-		return rawFileHandler(w, r, file)
-	}
-
-	return rawDirHandler(w, r, d, file)
-}
-
-func addFile(ar archiver.Writer, d *common.Data, path, commonPath string) error {
-	info, err := files.DefaultFs.Stat(path)
-	if err != nil {
-		return err
-	}
-
-	if !info.IsDir() && !info.Mode().IsRegular() {
-		return nil
-	}
-
-	file, err := files.DefaultFs.Open(path)
-	if err != nil {
-		return err
-	}
-	defer file.Close()
-
-	if path != commonPath {
-		filename := strings.TrimPrefix(path, commonPath)
-		filename = strings.TrimPrefix(filename, string(filepath.Separator))
-		err = ar.Write(archiver.File{
-			FileInfo: archiver.FileInfo{
-				FileInfo:   info,
-				CustomName: filename,
-			},
-			ReadCloser: file,
-		})
-		if err != nil {
-			return err
-		}
-	}
-
-	if info.IsDir() {
-		names, err := file.Readdirnames(0)
-		if err != nil {
-			return err
-		}
-
-		for _, name := range names {
-			fPath := filepath.Join(path, name)
-			err = addFile(ar, d, fPath, commonPath)
-			if err != nil {
-				klog.Errorf("Failed to archive %s: %v", fPath, err)
-			}
-		}
-	}
-
-	return nil
-}
-
-func rawDirHandler(w http.ResponseWriter, r *http.Request, d *common.Data, file *files.FileInfo) (int, error) {
-	filenames, err := parseQueryFiles(r, file)
-	if err != nil {
-		return http.StatusInternalServerError, err
-	}
-
-	extension, ar, err := parseQueryAlgorithm(r)
-	if err != nil {
-		return http.StatusInternalServerError, err
-	}
-
-	err = ar.Create(w)
-	if err != nil {
-		return http.StatusInternalServerError, err
-	}
-	defer ar.Close()
-
-	commonDir := fileutils.CommonPrefix(filepath.Separator, filenames...)
-
-	name := filepath.Base(commonDir)
-	if name == "." || name == "" || name == string(filepath.Separator) {
-		name = file.Name
-	}
-	// Prefix used to distinguish a filelist generated
-	// archive from the full directory archive
-	if len(filenames) > 1 {
-		name = "_" + name
-	}
-	name += extension
-	w.Header().Set("Content-Disposition", "attachment; filename*=utf-8''"+url.PathEscape(name))
-
-	for _, fname := range filenames {
-		err = addFile(ar, d, fname, commonDir)
-		if err != nil {
-			klog.Errorf("Failed to archive %s: %v", fname, err)
-		}
-	}
-
-	return 0, nil
-}
-
-func rawFileHandler(w http.ResponseWriter, r *http.Request, file *files.FileInfo) (int, error) {
-	fd, err := file.Fs.Open(file.Path)
-	if err != nil {
-		return http.StatusInternalServerError, err
-	}
-	defer fd.Close()
-
-	setContentDisposition(w, r, file)
-
-	w.Header().Set("Cache-Control", "private")
-	http.ServeContent(w, r, file.Name, file.ModTime, fd)
-	return 0, nil
+	return handler.RawHandler(w, r, d)
 }
